optimized code and add jaccard test
