+++ drain3/drain.py
@@ -49,7 +49,7 @@ class Node:
         self.cluster_ids: List[int] = []
 
 
-class Drain:
+class DrainBase:
     def __init__(self,
                  depth=4,
                  sim_th=0.4,
@@ -102,6 +102,144 @@ class Drain:
     @staticmethod
     def has_numbers(s):
         return any(char.isdigit() for char in s)
+
+    def fast_match(self, cluster_ids: Sequence, tokens: list, sim_th: float, include_params: bool):
+        """
+        Find the best match for a log message (represented as tokens) versus a list of clusters
+        :param cluster_ids: List of clusters to match against (represented by their IDs)
+        :param tokens: the log message, separated to tokens.
+        :param sim_th: minimum required similarity threshold (None will be returned in no clusters reached it)
+        :param include_params: consider tokens matched to wildcard parameters in similarity threshold.
+        :return: Best match cluster or None
+        """
+        match_cluster = None
+
+        max_sim = -1
+        max_param_count = -1
+        max_cluster = None
+
+        for cluster_id in cluster_ids:
+            # Try to retrieve cluster from cache with bypassing eviction
+            # algorithm as we are only testing candidates for a match.
+            cluster = self.id_to_cluster.get(cluster_id)
+            if cluster is None:
+                continue
+            cur_sim, param_count = self.get_seq_distance(cluster.log_template_tokens, tokens, include_params)
+            if cur_sim > max_sim or (cur_sim == max_sim and param_count > max_param_count):
+                max_sim = cur_sim
+                max_param_count = param_count
+                max_cluster = cluster
+
+        if max_sim >= sim_th:
+            match_cluster = max_cluster
+
+        return match_cluster
+
+    def print_tree(self, file=None, max_clusters=5):
+        self.print_node("root", self.root_node, 0, file, max_clusters)
+
+    def print_node(self, token, node, depth, file, max_clusters):
+        out_str = '\t' * depth
+
+        if depth == 0:
+            out_str += f'<{token}>'
+        elif depth == 1:
+            if token.isdigit():
+                out_str += f'<L={token}>'
+            else:
+                out_str += f'<{token}>'
+        else:
+            out_str += f'"{token}"'
+
+        if len(node.cluster_ids) > 0:
+            out_str += f" (cluster_count={len(node.cluster_ids)})"
+
+        print(out_str, file=file)
+
+        for token, child in node.key_to_child_node.items():
+            self.print_node(token, child, depth + 1, file, max_clusters)
+
+        for cid in node.cluster_ids[:max_clusters]:
+            cluster = self.id_to_cluster[cid]
+            out_str = '\t' * (depth + 1) + str(cluster)
+            print(out_str, file=file)
+
+    def get_content_as_tokens(self, content):
+        content = content.strip()
+        for delimiter in self.extra_delimiters:
+            content = content.replace(delimiter, " ")
+        content_tokens = content.split()
+        return content_tokens
+
+    def add_log_message(self, content: str):
+        content_tokens = self.get_content_as_tokens(content)
+
+        if self.profiler:
+            self.profiler.start_section("tree_search")
+        match_cluster = self.tree_search(self.root_node, content_tokens, self.sim_th, False)
+        if self.profiler:
+            self.profiler.end_section()
+
+        # Match no existing log cluster
+        if match_cluster is None:
+            if self.profiler:
+                self.profiler.start_section("create_cluster")
+            self.clusters_counter += 1
+            cluster_id = self.clusters_counter
+            match_cluster = LogCluster(content_tokens, cluster_id)
+            self.id_to_cluster[cluster_id] = match_cluster
+            self.add_seq_to_prefix_tree(self.root_node, match_cluster)
+            update_type = "cluster_created"
+
+        # Add the new log message to the existing cluster
+        else:
+            if self.profiler:
+                self.profiler.start_section("cluster_exist")
+            new_template_tokens = self.create_template(content_tokens, match_cluster.log_template_tokens)
+            if tuple(new_template_tokens) == match_cluster.log_template_tokens:
+                update_type = "none"
+            else:
+                match_cluster.log_template_tokens = tuple(new_template_tokens)
+                update_type = "cluster_template_changed"
+            match_cluster.size += 1
+            # Touch cluster to update its state in the cache.
+            # noinspection PyStatementEffect
+            self.id_to_cluster[match_cluster.cluster_id]
+
+        if self.profiler:
+            self.profiler.end_section()
+
+        return match_cluster, update_type
+
+    def get_total_cluster_size(self):
+        size = 0
+        for c in self.id_to_cluster.values():
+            size += c.size
+        return size
+
+    def get_clusters_ids_for_seq_len(self, seq_fir):
+        """
+        seq_fir: int/str - the first token of the sequence
+        Return all clusters with the specified count of tokens
+        """
+
+        def append_clusters_recursive(node: Node, id_list_to_fill: list):
+            id_list_to_fill.extend(node.cluster_ids)
+            for child_node in node.key_to_child_node.values():
+                append_clusters_recursive(child_node, id_list_to_fill)
+
+        cur_node = self.root_node.key_to_child_node.get(str(seq_fir))
+
+        # no template with same token count
+        if cur_node is None:
+            return []
+
+        target = []
+        append_clusters_recursive(cur_node, target)
+        return target
+
+
+class Drain(DrainBase):
 
     def tree_search(self, root_node: Node, tokens: list, sim_th: float, include_params: bool):
 
@@ -232,38 +370,6 @@ class Drain:
 
         return ret_val, param_count
 
-    def fast_match(self, cluster_ids: Sequence, tokens: list, sim_th: float, include_params: bool):
-        """
-        Find the best match for a log message (represented as tokens) versus a list of clusters
-        :param cluster_ids: List of clusters to match against (represented by their IDs)
-        :param tokens: the log message, separated to tokens.
-        :param sim_th: minimum required similarity threshold (None will be returned in no clusters reached it)
-        :param include_params: consider tokens matched to wildcard parameters in similarity threshold.
-        :return: Best match cluster or None
-        """
-        match_cluster = None
-
-        max_sim = -1
-        max_param_count = -1
-        max_cluster = None
-
-        for cluster_id in cluster_ids:
-            # Try to retrieve cluster from cache with bypassing eviction
-            # algorithm as we are only testing candidates for a match.
-            cluster = self.id_to_cluster.get(cluster_id)
-            if cluster is None:
-                continue
-            cur_sim, param_count = self.get_seq_distance(cluster.log_template_tokens, tokens, include_params)
-            if cur_sim > max_sim or (cur_sim == max_sim and param_count > max_param_count):
-                max_sim = cur_sim
-                max_param_count = param_count
-                max_cluster = cluster
-
-        if max_sim >= sim_th:
-            match_cluster = max_cluster
-
-        return match_cluster
-
     def create_template(self, seq1, seq2):
         assert len(seq1) == len(seq2)
         ret_val = list(seq2)
@@ -273,99 +379,6 @@ class Drain:
                 ret_val[i] = self.param_str
 
         return ret_val
-
-    def print_tree(self, file=None, max_clusters=5):
-        self.print_node("root", self.root_node, 0, file, max_clusters)
-
-    def print_node(self, token, node, depth, file, max_clusters):
-        out_str = '\t' * depth
-
-        if depth == 0:
-            out_str += f'<{token}>'
-        elif depth == 1:
-            out_str += f'<L={token}>'
-        else:
-            out_str += f'"{token}"'
-
-        if len(node.cluster_ids) > 0:
-            out_str += f" (cluster_count={len(node.cluster_ids)})"
-
-        print(out_str, file=file)
-
-        for token, child in node.key_to_child_node.items():
-            self.print_node(token, child, depth + 1, file, max_clusters)
-
-        for cid in node.cluster_ids[:max_clusters]:
-            cluster = self.id_to_cluster[cid]
-            out_str = '\t' * (depth + 1) + str(cluster)
-            print(out_str, file=file)
-
-    def get_content_as_tokens(self, content):
-        content = content.strip()
-        for delimiter in self.extra_delimiters:
-            content = content.replace(delimiter, " ")
-        content_tokens = content.split()
-        return content_tokens
-
-    def add_log_message(self, content: str):
-        content_tokens = self.get_content_as_tokens(content)
-
-        if self.profiler:
-            self.profiler.start_section("tree_search")
-        match_cluster = self.tree_search(self.root_node, content_tokens, self.sim_th, False)
-        if self.profiler:
-            self.profiler.end_section()
-
-        # Match no existing log cluster
-        if match_cluster is None:
-            if self.profiler:
-                self.profiler.start_section("create_cluster")
-            self.clusters_counter += 1
-            cluster_id = self.clusters_counter
-            match_cluster = LogCluster(content_tokens, cluster_id)
-            self.id_to_cluster[cluster_id] = match_cluster
-            self.add_seq_to_prefix_tree(self.root_node, match_cluster)
-            update_type = "cluster_created"
-
-        # Add the new log message to the existing cluster
-        else:
-            if self.profiler:
-                self.profiler.start_section("cluster_exist")
-            new_template_tokens = self.create_template(content_tokens, match_cluster.log_template_tokens)
-            if tuple(new_template_tokens) == match_cluster.log_template_tokens:
-                update_type = "none"
-            else:
-                match_cluster.log_template_tokens = tuple(new_template_tokens)
-                update_type = "cluster_template_changed"
-            match_cluster.size += 1
-            # Touch cluster to update its state in the cache.
-            # noinspection PyStatementEffect
-            self.id_to_cluster[match_cluster.cluster_id]
-
-        if self.profiler:
-            self.profiler.end_section()
-
-        return match_cluster, update_type
-
-    def get_clusters_ids_for_seq_len(self, seq_len: int):
-        """
-        Return all clusters with the specified count of tokens
-        """
-
-        def append_clusters_recursive(node: Node, id_list_to_fill: list):
-            id_list_to_fill.extend(node.cluster_ids)
-            for child_node in node.key_to_child_node.values():
-                append_clusters_recursive(child_node, id_list_to_fill)
-
-        cur_node = self.root_node.key_to_child_node.get(str(seq_len))
-
-        # no template with same token count
-        if cur_node is None:
-            return []
-
-        target = []
-        append_clusters_recursive(cur_node, target)
-        return target
 
     def match(self, content: str, full_search_strategy="never"):
         """
@@ -399,240 +412,6 @@ class Drain:
         # quitting on less than exact cluster matches.
         def full_search():
             all_ids = self.get_clusters_ids_for_seq_len(len(content_tokens))
-            cluster = self.fast_match(all_ids, content_tokens, required_sim_th, include_params=True)
-            return cluster
-
-        if full_search_strategy == "always":
-            return full_search()
-
-        match_cluster = self.tree_search(self.root_node, content_tokens, required_sim_th, include_params=True)
-        if match_cluster is not None:
-            return match_cluster
-
-        if full_search_strategy == "never":
-            return None
-
-        return full_search()
-
-    def get_total_cluster_size(self):
-        size = 0
-        for c in self.id_to_cluster.values():
-            size += c.size
-        return size
-
-
-class JaccardDrain(Drain):
-    """
-    add a new matching pattern to the log cluster.
-    Cancels log message length as  first token.
-    Drain that uses Jaccard similarity to match log messages.
-    """
-
-    def tree_search(self, root_node: Node, tokens: list, sim_th: float, include_params: bool):
-
-        # at first level, children are grouped by token (The first word in tokens)
-        token_count = len(tokens)
-
-        if not tokens:
-            token_first = ""
-            cur_node = root_node.key_to_child_node.get(token_first)
-        else:
-            token_first = tokens[0]
-            cur_node = root_node.key_to_child_node.get(token_first)
-
-        # no template with same token count yet
-        if cur_node is None:
-            return None
-
-        # handle case of empty log string - return the single cluster in that group
-        if token_count == 0:
-            return self.id_to_cluster.get(cur_node.cluster_ids[0])
-
-        # find the leaf node for this log - a path of nodes matching the first N tokens (N=tree depth)
-        cur_node_depth = 1   # first level is 1 <root>
-
-        for token in tokens[1:]:
-            # at max depth
-            if cur_node_depth >= self.max_node_depth:
-                break
-
-            # this is last token
-            # It starts with the second word, so the sentence length -1
-            if cur_node_depth == token_count - 1:
-                break
-
-            key_to_child_node = cur_node.key_to_child_node
-            cur_node = key_to_child_node.get(token)
-
-            if cur_node is None:  # no exact next token exist, try wildcard node
-                cur_node = key_to_child_node.get(self.param_str)
-            if cur_node is None:  # no wildcard node exist
-                return None
-
-            cur_node_depth += 1
-
-        # get best match among all clusters with same prefix, or None if no match is above sim_th
-        cluster = self.fast_match(cur_node.cluster_ids, tokens, sim_th, include_params)
-
-        return cluster
-
-    def add_seq_to_prefix_tree(self, root_node, cluster: LogCluster):
-        token_count = len(cluster.log_template_tokens)
-        # Determine if the string is empty
-        if not cluster.log_template_tokens:
-            token_first = ""
-        else:
-            token_first = cluster.log_template_tokens[0]
-        if token_first not in root_node.key_to_child_node:
-            first_layer_node = Node()
-            root_node.key_to_child_node[token_first] = first_layer_node
-        else:
-            first_layer_node = root_node.key_to_child_node[token_first]
-
-        cur_node = first_layer_node
-
-        # handle case of empty log string
-        if token_count == 0:
-            cur_node.cluster_ids = [cluster.cluster_id]
-            return
-
-        current_depth = 1
-        for token in cluster.log_template_tokens[1:]:
-
-            # if at max depth or this is last token in template - add current log cluster to the leaf node
-            # It starts with the second word, so the sentence length -1
-            if current_depth >= self.max_node_depth or current_depth >= token_count - 1:
-                # clean up stale clusters before adding a new one.
-                new_cluster_ids = []
-                for cluster_id in cur_node.cluster_ids:
-                    if cluster_id in self.id_to_cluster:
-                        new_cluster_ids.append(cluster_id)
-                new_cluster_ids.append(cluster.cluster_id)
-                cur_node.cluster_ids = new_cluster_ids
-                break
-
-            # if token not matched in this layer of existing tree.
-            if token not in cur_node.key_to_child_node:
-                if self.parametrize_numeric_tokens and self.has_numbers(token):
-                    if self.param_str not in cur_node.key_to_child_node:
-                        new_node = Node()
-                        cur_node.key_to_child_node[self.param_str] = new_node
-                        cur_node = new_node
-                    else:
-                        cur_node = cur_node.key_to_child_node[self.param_str]
-
-                else:
-                    if self.param_str in cur_node.key_to_child_node:
-                        if len(cur_node.key_to_child_node) < self.max_children:
-                            new_node = Node()
-                            cur_node.key_to_child_node[token] = new_node
-                            cur_node = new_node
-                        else:
-                            cur_node = cur_node.key_to_child_node[self.param_str]
-                    else:
-                        if len(cur_node.key_to_child_node) + 1 < self.max_children:
-                            new_node = Node()
-                            cur_node.key_to_child_node[token] = new_node
-                            cur_node = new_node
-                        elif len(cur_node.key_to_child_node) + 1 == self.max_children:
-                            new_node = Node()
-                            cur_node.key_to_child_node[self.param_str] = new_node
-                            cur_node = new_node
-                        else:
-                            cur_node = cur_node.key_to_child_node[self.param_str]
-
-            # if the token is matched
-            else:
-                cur_node = cur_node.key_to_child_node[token]
-
-            current_depth += 1
-
-    # seq1 is a template, seq2 is the log to match
-    def get_seq_distance(self, seq1: tuple, seq2: list, include_params: bool):
-        # Jaccard index, It is used to measure the similarity of two sets.
-        # The closer its value is to 1, the more common members the two sets have, and the higher the similarity.
-
-        # sequences are empty - full match
-        if len(seq1) == 0:
-            return 1.0, 0
-
-        param_count = 0
-
-        for token1 in seq1:
-            if token1 == self.param_str:
-                param_count += 1
-
-        # If there are param_str, they are removed from the coefficient calculation
-        if include_params:
-            seq1 = (x for x in seq1 if x != self.param_str)
-
-        # Calculate the Jaccard coefficient
-        ret_val = len(set(seq1) & set(seq2)) / len(set(seq1) | set(seq2))
-
-        return ret_val, param_count
-
-    # seq1:tonkens->list seq2:template->tuple
-    def create_template(self, seq1:list, seq2: tuple):
-
-        inter_set = set(seq1) & set(seq2)
-        # Take the template with long length
-        ret_val = seq1 if len(seq1) > len(seq2) else list(seq2)
-
-        for i, token in enumerate(ret_val):
-            if token not in inter_set:
-                ret_val[i] = self.param_str
-
-        return ret_val
-
-    def print_node(self, token, node, depth, file, max_clusters):
-        out_str = '\t' * depth
-
-        if depth == 0:
-            out_str += f'<{token}>'
-        elif depth == 1:
-            out_str += f'<{token}>'
-        else:
-            out_str += f'"{token}"'
-
-        if len(node.cluster_ids) > 0:
-            out_str += f" (cluster_count={len(node.cluster_ids)})"
-
-        print(out_str, file=file)
-
-        for token, child in node.key_to_child_node.items():
-            self.print_node(token, child, depth + 1, file, max_clusters)
-
-        for cid in node.cluster_ids[:max_clusters]:
-            cluster = self.id_to_cluster[cid]
-            out_str = '\t' * (depth + 1) + str(cluster)
-            print(out_str, file=file)
-
-    def get_clusters_ids_for_seq_len(self, seq_first):
-
-        def append_clusters_recursive(node: Node, id_list_to_fill: list):
-            id_list_to_fill.extend(node.cluster_ids)
-            for child_node in node.key_to_child_node.values():
-                append_clusters_recursive(child_node, id_list_to_fill)
-
-        cur_node = self.root_node.key_to_child_node.get(seq_first)
-
-        # no template with same token first word
-        if cur_node is None:
-            return []
-
-        target = []
-        append_clusters_recursive(cur_node, target)
-        return target
-
-    def match(self, content: str, full_search_strategy="never"):
-
-        assert full_search_strategy in ["always", "never", "fallback"]
-
-        required_sim_th = 1.0
-        content_tokens = self.get_content_as_tokens(content)
-
-        def full_search():
-            all_ids = self.get_clusters_ids_for_seq_len(content_tokens[0])
             cluster = self.fast_match(all_ids, content_tokens, required_sim_th, include_params=True)
             return cluster
