+++ tests/drain3_test.ini
@@ -14,6 +14,7 @@ masking = [
           ]
 
 [DRAIN]
+engine = Drain
 sim_th = 0.4
 depth = 4
 max_children = 100
